moving read static modules
