+++ packages/amplication-data-service-generator/src/util/read-static-modules.ts
